Split adapters and adapter manager off index, move files around
