+++ packages/jupyterlab-lsp/src/editor_integration/cm_adapter.spec.ts
@@ -1,11 +1,11 @@
 import { expect } from 'chai';
-import { LSPConnection } from '../../connection';
-import { StatusMessage } from '../jupyterlab/jl_adapter';
-import { IRootPosition } from '../../positioning';
+import { LSPConnection } from '../connection';
+import { StatusMessage } from '../adapters/adapter';
+import { IRootPosition } from '../positioning';
 import * as CodeMirror from 'codemirror';
 import { FileEditorFeatureTestEnvironment } from './testutils';
-import { CodeMirrorIntegration } from '../../editor_integration/codemirror';
-import { EditorAdapter } from '../editor_adapter';
+import { CodeMirrorIntegration } from './codemirror';
+import { EditorAdapter } from './editor_adapter';
 
 describe('CodeMirrorAdapter', () => {
   let env: FileEditorFeatureTestEnvironment;
